Fixed the type_ids.md document
The byte format positions and lens for the PEER_INFO and MY_INFO
type ids were fixed to properly reflect format(s).

Few format changes to the code using the 'auto' keyword. This will
be an ongoing thing on every commit until all parts of the code
use the 'auto' keyword were appropriate.