Bug fixes for Windows
fixed a bug in windows that caused the application to
fail detect that there is a host instance running
when there is indeed a host instance running.

also fixed a compile bug that cause QtCore to double
import and in turn call out duplicate definition
errors in windows when including applink.c.